Extract RequestDownloadDialogs into its own component
